+++ R/plots.R
@@ -7,7 +7,7 @@
 #' @param var_time variable that measures time.
 #' @param n number of observation per id.
 #' @param t value of the confidence interval.
-#'
+#' @param threshold distance in kilometers. By default is 800.
 #'
 #' @return A ggplot object
 #'
@@ -37,9 +37,10 @@ plot_Moran.I <- function(data,
                          var_id = character(),
                          var_time = character(),
                          n = 1,
-                         t = 1.645){
+                         t = 1.645,
+                         threshold = 800L){
 
-    wdata <- data_plots(data = data, var_id = var_id, var_time = var_time, n = n)
+    wdata <- data_plots(data = data, var_id = var_id, var_time = var_time, n = n, threshold = threshold)
     w2   <- wdata[[2]]
     mats <- wdata[[1]]
     t_ <- t
@@ -68,7 +69,7 @@ plot_Moran.I <- function(data,
 }
 
 
-#' @title plot_JointCount
+#' @title plot_JoinCount
 #' @description Uses Joint Count tests to assess spatial clustering or dispersion of categorical variables in the data. Negative values indicate positive spatial clustering.
 #'
 #' @param data data.
@@ -77,6 +78,7 @@ plot_Moran.I <- function(data,
 #' @param var_time variable that measures time.
 #' @param n number of observation per id.
 #' @param t value of the confidence interval.
+#' @param threshold distance in kilometers. By default is 800.
 #'
 #' @return A ggplot object
 #'
@@ -91,7 +93,7 @@ plot_Moran.I <- function(data,
 #'                                    ongoing = FALSE)
 #'
 #'
-#' plot_JointCount(data = dataw,
+#' plot_JoinCount(data = dataw,
 #'                var_cured = "cured",
 #'                var_id = "ccode",
 #'                var_time = "year",
@@ -100,15 +102,16 @@ plot_Moran.I <- function(data,
 #'
 #' @export
 
-plot_JointCount <- function(data,
+plot_JoinCount <- function(data,
                            var_cured = character(),
                            var_id = character(),
                            var_time = character(),
                            n = 1,
-                           t = 1.645){
+                           t = 1.645,
+                           threshold = 800L){
 
 
-    wdata <- data_plots(data = data, var_id = var_id, var_time = var_time, n = n)
+    wdata <- data_plots(data = data, var_id = var_id, var_time = var_time, n = n, threshold = threshold)
     w2   <- wdata[[2]]
     mats <- wdata[[1]]
     t_ <- t
@@ -139,5 +142,3 @@ plot_JointCount <- function(data,
         ggplot2::theme_bw() +
         ggplot2::theme(axis.text = ggplot2::element_text(size = 10), axis.title = ggplot2::element_text(size=15, face = "bold"))
 }
-
-