+++ NEWS.md
@@ -1,3 +1,14 @@
+
+# BayesSPsurv 0.1.2
+
+* threshold argument added to `plot_Moran.I()` and `plot_JoinCount()`
+* parameter argument added to generic function `plot()`
+* rename plot_JointCount(): `plot_JoinCount()`
+
 # BayesSPsurv 0.1.1
 
 * add `plot_Moran.I()` and `plot_JointCount()`
+
+
+
+